Various

+++ core/src/main/java/com/boydti/fawe/object/ChangeSetFaweQueue.java
@@ -23,7 +23,7 @@ public class ChangeSetFaweQueue extends DelegateFaweQueue {
     }
 
     @Override
-    public boolean setBlock(int x, int y, int z, short id, byte data) {
+    public boolean setBlock(int x, int y, int z, int id, int data) {
         if (super.setBlock(x, y, z, id, data)) {
             int combinedFrom = getParent().getCombinedId4Data(x, y, z);
             if (FaweCache.hasNBT(combinedFrom >> 4)) {
